requirements.txt
appdirs==1.4.4
asgiref==3.2.7
boto==2.49.0
boto3==1.13.19
botocore==1.16.19
bson==0.5.8
certifi==2020.4.5.1
chardet==3.0.4
dataclasses==0.6
distlib==0.3.0
Django==3.0.6
django-boto==0.3.12
django-rest-framework-mongoengine==3.4.1
django-storages==1.9.1
djangorestframework==3.11.0
djongo==1.3.2
dnspython==1.16.0
docutils==0.15.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
jmespath==0.10.0
mongoengine==0.20.0
mysqlclient==1.4.6
numpy==1.18.4
Pillow==7.1.2
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.2.4
Unipath==1.1
urllib3==1.25.9
virtualenv==20.0.21
zipp==3.1.0
